Updates to /utils structure
