+++ utils/__init__.py
@@ -1,3 +1,1 @@
-from .utils import get_ticker, Embedding, Interaction
-from .sentiment.headlines import get_headlines
-from .sentiment.social_media import get_social_media
+from .utils import get_ticker
